feat: add state updating functionality for task
This is the first part of manipulating data for a specific task.
When a subtask of a task is toggled in the modal, the toggleCompleted
func is invoked and the task is updated. This func in turn updates the
task state of the ViewTask component and invokes the updateTask
of the boardctx.js context file to update the boards state.