fix: update the ids state when changing status
Prior to this fix, whenever the status for a task was changed
the modal would not be rendered even though the state
controlling it showed it should be. After some digging around
I found that I wasn't updating the column id key of the ids state
which rendered the modal. This commit implements that change.